#1 MySQL support 20

+++ src/main/java/me/tabinol/secuboid/storage/mysql/dao/ApprovesDao.java
@@ -49,10 +49,10 @@ public final class ApprovesDao {
             try (final ResultSet rs = stmt.executeQuery()) {
                 while (rs.next()) {
                     final UUID landUUID = DbUtils.getUUID(rs, "land_uuid");
-                    final int approveActionId = rs.getInt("approve_action_id");
+                    final long approveActionId = rs.getLong("approve_action_id");
                     final Optional<Integer> removedAreaIdOpt = DbUtils.getOpt(rs, "removed_area_id", rs::getInt);
                     final Optional<Integer> newAreaIdOpt = DbUtils.getOpt(rs, "new_area_id", rs::getInt);
-                    final int ownerId = rs.getInt("owner_id");
+                    final long ownerId = rs.getLong("owner_id");
                     final Optional<UUID> parentUUIDOpt = DbUtils.getOpt(rs, "parent_uuid", c -> DbUtils.getUUID(rs, c));
                     final double price = rs.getDouble("price");
                     Calendar transactionDatetime = DbUtils.getCalendar(rs, "transaction_datetime");
@@ -76,18 +76,18 @@ public final class ApprovesDao {
 
         try (final PreparedStatement stmt = dbConn.preparedStatementWithTags(conn, sql)) {
             DbUtils.setUUID(stmt, 1, approvePojo.getLandUUID());
-            stmt.setInt(2, approvePojo.getApproveActionId());
+            stmt.setLong(2, approvePojo.getApproveActionId());
             DbUtils.setOpt(stmt, 3, approvePojo.getRemovedAreaIdOpt(), (i, u) -> stmt.setInt(i, u));
             DbUtils.setOpt(stmt, 4, approvePojo.getNewAreaIdOpt(), (i, u) -> stmt.setInt(i, u));
-            stmt.setInt(5, approvePojo.getOwnerId());
+            stmt.setLong(5, approvePojo.getOwnerId());
             DbUtils.setOpt(stmt, 6, approvePojo.getParentUUIDOpt(), (i, u) -> DbUtils.setUUID(stmt, i, u));
             stmt.setDouble(7, approvePojo.getPrice());
             DbUtils.setCalendar(stmt, 8, approvePojo.getTransactionDatetime());
 
-            stmt.setInt(9, approvePojo.getApproveActionId());
+            stmt.setLong(9, approvePojo.getApproveActionId());
             DbUtils.setOpt(stmt, 10, approvePojo.getRemovedAreaIdOpt(), (i, u) -> stmt.setInt(i, u));
             DbUtils.setOpt(stmt, 11, approvePojo.getNewAreaIdOpt(), (i, u) -> stmt.setInt(i, u));
-            stmt.setInt(12, approvePojo.getOwnerId());
+            stmt.setLong(12, approvePojo.getOwnerId());
             DbUtils.setOpt(stmt, 13, approvePojo.getParentUUIDOpt(), (i, u) -> DbUtils.setUUID(stmt, i, u));
             stmt.setDouble(14, approvePojo.getPrice());
             DbUtils.setCalendar(stmt, 15, approvePojo.getTransactionDatetime());
